requirements.txt
Flask==2.3.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
Flask-Admin==1.6.0
python-dotenv==1.0.0
razorpay==1.3.0
Werkzeug==2.3.7
